% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{response_schema}
\alias{response_schema}
\title{Esquema JSON de respuesta esperada (response schema para Gemini)}
\usage{
response_schema()
}
\description{
Esquema JSON de respuesta esperada (response schema para Gemini)
}
\keyword{internal}
